fix:TOP-58 type 수정으로 인한 수정

+++ feature/tohot/src/main/java/tht/feature/tohot/component/topic/TopicSelectChip.kt
@@ -29,10 +29,10 @@ fun TopicSelectChip(
     imageUrl: String?,
     @DrawableRes imageRes: Int,
     title: String,
-    key: Long,
+    key: Int,
     content: String,
     isSelect: Boolean,
-    topicClickListener: (Long) -> Unit = { }
+    topicClickListener: (Int) -> Unit = { }
 ) {
     Row(
         modifier = modifier
